laravel-project

+++ event-management/app/Http/Controllers/Api/EventController.php
@@ -7,6 +7,7 @@ use App\Http\Resources\EventResource;
 use App\Http\Traits\CanLoadRelationships;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class EventController extends Controller
 {
@@ -22,6 +23,9 @@ class EventController extends Controller
      public function __construct()
      {
         $this->middleware('auth:sanctum')->except(['index','show']);
+        $this->middleware('throttle:60,1')->only(['store','destroy']);
+        $this->authorizeResource(Event::class,'event');
+       
      }
 
     public function index()
@@ -64,6 +68,10 @@ class EventController extends Controller
     public function update(Request $request, Event $event)
     {
 
+        // if(Gate::denies('update-event',$event)){
+        //   abort(403,'You are Not authorized to update this event');
+        // }
+        // $this->authorize('update-event',$event);
         $event->update([
             ...$request->validate([
                 'name' => 'sometimes|string|max:255',
@@ -81,6 +89,7 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
+        $this->authorize('delete-event',$event);
         $event->delete();
 
         return response('',204);
